Create separate UI scenario in molecule. Revert travis.yml back the way it was. Run Jenkins pipeline against a single OS.
