--- .travis.yml.bak
@@ -1,66 +1,0 @@
----
-
-branches:
-  only:
-    - "master"
-
-dist: focal
-
-language: python
-python: 3.8
-
-addons:
-  chrome: stable
-  apt:
-    packages:
-      - python3-pip
-      - libssl-dev
-      - python3-virtualenv
-      - ruby-libvirt
-      - qemu
-      - libvirt-daemon-system
-      - libvirt-clients
-      - libxslt-dev
-      - libxml2-dev
-      - libvirt-dev
-      - zlib1g-dev
-      - bridge-utils
-      - dnsmasq-base
-      - ebtables
-      - libvirt-dev
-      - qemu-kvm
-      - qemu-utils
-      - ruby-dev
-      - chromium-chromedriver
-
-
-env:
-  matrix:
-    - MOLECULE_DISTRO: centos/7
-    - MOLECULE_DISTRO: centos/8
-    - MOLECULE_DISTRO: debian/stretch64
-    - MOLECULE_DISTRO: debian/buster64
-    - MOLECULE_DISTRO: generic/ubuntu1604
-    - MOLECULE_DISTRO: generic/ubuntu1804
-    - MOLECULE_DISTRO: generic/ubuntu2004
-
-before_install:
-  - sudo apt-get update
-
-install:
-  #Install Vagrant
-  - |
-    export VAGRANT_VERSION=2.2.10
-    sudo wget -nv https://releases.hashicorp.com/vagrant/${VAGRANT_VERSION}/vagrant_${VAGRANT_VERSION}_x86_64.deb
-    sudo dpkg -i vagrant_${VAGRANT_VERSION}_x86_64.deb
-    vagrant --version
-  - sudo vagrant plugin install vagrant-libvirt
-  - pip3 install molecule yamllint ansible-lint python-vagrant molecule-vagrant "molecule[lint]" testinfra selenium
-  - sudo chown -R travis:travis /home/travis/.vagrant.d
-  - sudo gpasswd -a travis libvirt
-
-before_script:
-  - ln -s /home/travis/bin/chromedriver ~/bin/chromedriver
-
-script:
-  - sudo -E su travis travis.sh
