+++ molecule/default/molecule.yml
@@ -1,3 +1,4 @@
+
 ---
 dependency:
   name: galaxy
@@ -7,18 +8,25 @@ dependency:
 driver:
   name: vagrant
   provider:
-    name: virtualbox
-lint: |
-  set -e
-  yamllint .
-  #ansible-lint
+    name: libvirt
 platforms:
   - name: instance
-    box: ${MOLECULE_DISTRO:-generic/ubuntu2004}
+    box: ${MOLECULE_DISTRO:-generic/ubuntu1804}
     memory: 4096
     cpus: 1
-    instance_raw_config_args:
-      - "vm.network 'forwarded_port', guest: 9000, host: 9000"
+    interfaces:
+      - auto_config: true
+        network_name: private_network
+        type: dhcp
+    provider_raw_config_args:
+      - 'nic_model_type = "virtio"'
+      - 'disk_bus = "virtio"'
+      - 'volume_cache = "none"'
+      - 'storage :file, bus: "virtio", cache: "none"'
+providers:
+  - name: libvirt
+    type: libvirt
+    driver: kvm
 provisioner:
   name: ansible
 verifier:
