+++ molecule/ui/converge.yml
@@ -1,0 +1,66 @@
+---
+- name: Converge
+  hosts: all
+  vars:
+    #Graylog
+    graylog_install_elasticsearch: True
+    graylog_install_mongodb:       True
+    graylog_install_nginx:         False
+    graylog_install_java:          True
+    graylog_not_testing:           False
+    graylog_version:               3.3
+    graylog_mongodb_version:       4.2
+    graylog_http_bind_address:     "{{ hostvars[inventory_hostname]['ansible_default_ipv4']['address'] }}:9000"
+    graylog_http_publish_uri:      "http://localhost:9000/"
+    graylog_http_external_uri:     "http://localhost:9000/"
+
+    #Elasticsearch
+    es_major_version: "6.x"
+    es_version: 6.8.10
+    es_instance_name: "graylog"
+    es_scripts: False
+    es_templates: False
+    es_version_lock: False
+    es_heap_size: "1g"
+    es_java_install: False
+    oss_version: True
+
+    es_config:
+      node.name: "graylog"
+      cluster.name: "graylog"
+      http.port: 9200
+      transport.tcp.port: 9300
+      network.host: "0.0.0.0"
+      node.data: True
+      node.master: True
+
+    #Java
+    java_distribution: adoptopenjdk
+    java_major_version: 8
+
+    nginx_sites:
+      graylog:
+        - "listen 80"
+        - "server_name graylog"
+        - |
+          location / {
+            proxy_pass http://localhost:9000/;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_pass_request_headers on;
+            proxy_connect_timeout 150;
+            proxy_send_timeout 100;
+            proxy_read_timeout 100;
+            proxy_buffers 4 32k;
+            client_max_body_size 8m;
+            client_body_buffer_size 128k;
+          }
+
+  tasks:
+    - name: "Include graylog-ansible-role"
+      include_role:
+        name: "graylog-ansible-role"
+    - debug:
+        msg: "Graylog is up at {{ graylog_http_external_uri }}"
+  become: yes
